Config moved to module
