+++ lib/coupon.js
@@ -1,0 +1,7 @@
+
+
+module.exports = {
+    create: function() {
+        console.log(process.nodester);
+    }
+}
